Split routes/agents.ts into domain-focused modules (#tech-debt)
The monolithic agents.ts (1212 lines, single registerAgentRoutes function)
is now split into routes/agents/ with focused modules: crud-routes.ts,
lifecycle-routes.ts, streaming-routes.ts, terminal-routes.ts,
events-routes.ts, types.ts, and an index.ts that composes them all.
